update: using bootstrap4

+++ utils.py
@@ -76,13 +76,13 @@ def cp(src,dest):
     print("copying file:" + src +" " + dest)
     if os.path.exists(dest):
         os.remove(dest)
-    shutil.copy(src,dest)
+    if os.path.exists(src):shutil.copy(src,dest)
 
 def cp_r(src,dest):
     print("copying folder:" + src +" " + dest)
     if os.path.exists(dest):
         shutil.rmtree(dest)
-    shutil.copytree(src,dest)
+    if os.path.exists(src):shutil.copytree(src,dest)
 
 def sha1(value):
     return hashlib.sha1(value.encode()).hexdigest()
